CNMining/CNMining.jar.src/uk/ac/shef/wit/simmetrics/math/MathFuncs.java
/*     */ package uk.ac.shef.wit.simmetrics.math;
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ public final class MathFuncs
/*     */ {
/*     */   public static float max3(float x, float y, float z)
/*     */   {
/*  61 */     return Math.max(x, Math.max(y, z));
/*     */   }
/*     */   
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */   public static float max4(float w, float x, float y, float z)
/*     */   {
/*  74 */     return Math.max(Math.max(w, x), Math.max(y, z));
/*     */   }
/*     */   
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */   public static float min3(float x, float y, float z)
/*     */   {
/*  86 */     return Math.min(x, Math.min(y, z));
/*     */   }
/*     */   
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */   public static int min3(int x, int y, int z)
/*     */   {
/*  98 */     return Math.min(x, Math.min(y, z));
/*     */   }
/*     */   
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */ 
/*     */   public static int max3(int x, int y, int z)
/*     */   {
/* 110 */     return Math.max(x, Math.max(y, z));
/*     */   }
/*     */ }


/* Location:              /home/menick/Scrivania/ProcessMining/CNMining/CNMining.jar!/uk/ac/shef/wit/simmetrics/math/MathFuncs.class
 * Java compiler version: 5 (49.0)
 * JD-Core Version:       0.7.1
 */